chore: move abstract socket class
