+++ src/sockets/input/VaryingInputSocket.ts
@@ -1,6 +1,6 @@
 import {INode} from '../../node/INode';
 import {ShaderVaryingObject} from '../../types/CommonType';
-import AbstractVaryingSocket from '../AbstractVaryingSocket';
+import AbstractVaryingSocket from '../abstract/AbstractVaryingSocket';
 import {IVaryingOutputSocket} from '../output/IVaryingOutputSocket';
 import {IVaryingInputSocket} from './IVaryingInputSocket';
